Use float for timing in particle system
Also, update the assets_loader due to this

+++ engine/particle_sys.h
@@ -22,7 +22,7 @@ typedef struct Particle
     float rotation;
     float angular_vel;
     float size;
-    uint32_t timer;
+    float timer;
     bool alive;
     bool spawned;
 }Particle_t;
@@ -38,8 +38,8 @@ typedef struct EmitterConfig
     float speed_range[2];
     float angle_range[2];
     float rotation_range[2];
-    uint32_t particle_lifetime[2];
-    uint32_t initial_spawn_delay;
+    float particle_lifetime[2];
+    float initial_spawn_delay;
     PartEmitterType_t type;
     bool one_shot;
 }EmitterConfig_t;
@@ -51,7 +51,7 @@ struct ParticleEmitter
     Vector2 position;
     Particle_t particles[MAX_PARTICLES];
     uint32_t n_particles;
-    uint32_t timer;
+    float timer;
     bool finished;
     bool active;
     void* user_data;
